+++ R/mse.R
@@ -6,9 +6,17 @@
 #' @param running logical, denoting whether to use the running fitted values.
 #' @param ... Other arguments (not in use at this time).
 #' @details Even though logistic regressions don't use the MSE on the y=0/1 scale, it can still be informative.
+#'   Note that the S3 method is \code{mse}.
 #' @name elo.mse
 NULL
 #> NULL
+
+#' @rdname elo.mse
+#' @export
+brier <- function(object, subset, ...)
+{
+  UseMethod("mse")
+}
 
 #' @rdname elo.mse
 #' @export